+++ string_calculator/run.sh
@@ -1,5 +1,5 @@
-FILE=./main/target/main-0.2.0-jar-with-dependencies.jar
-if [ ! -f "$FILE" ]; then
+#FILE=./main/target/main-0.2.0-jar-with-dependencies.jar
+#if [ ! -f "$FILE" ]; then
   mvn package
-fi
+#fi
 java -jar main/target/main-0.2.0-jar-with-dependencies.jar